pyproject.toml
[project]
name = "umap-autoencoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.31.0",
    "ipywidgets>=8.1.8",
    "matplotlib>=3.9.4",
    "mediapipe>=0.10.14",
    "numpy>=2.0.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "torch>=2.8.0",
    "umap-learn>=0.5.9.post2",
]
